Update to file structure, may need optimization (probably needs optimization)
