+++ .babelrc
@@ -1,7 +1,7 @@
 {
-    "presets": [
-        "@babel/react",
-        "@babel/env",
-        "@babel/preset-typescript"
-    ]
+    "env": {
+      "development" : {
+        "compact": false
+      }
+    }
 }
