+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 build/
+dist/
 .firebase/
 package-lock.json
 yarn.lock
