+++ README.md
@@ -1,6 +1,9 @@
 # react-world-countries-map
 A pure react component to visualize world maps with highlighted countries. Simple. Free. No Registration, No access token, No back-end service.
 
+# WORK IN PROGRESS, does not import properly yet
+
+
 ~~~
 import ReactWorldCountriesMap from "react-world-countries-map"
 ...
