+++ example/App.css
@@ -1,0 +1,11 @@
+.App {
+  text-align: center;
+  background: black;
+  color: white;
+  height: 100vh;
+  width: 100vw;
+}
+
+.Main {
+  padding: 10px;
+}
