--- examples/src/App.css
@@ -1,11 +1,0 @@
-.App {
-  text-align: center;
-  background: black;
-  color: white;
-  height: 100vh;
-  width: 100vw;
-}
-
-.Main {
-  padding: 10px;
-}
