--- src/ReactWorldCountriesMap.css
@@ -1,10 +1,0 @@
-.mapView {
-    height: auto;
-    width: auto;
-    padding: 0px;
-    margin: 0px;
-}
-
-.map {
-    background-color: #ffffff;
-}
